Refactor to manage dependencies
