+++ st2common/st2common/services/action.py
@@ -54,7 +54,7 @@ def create_request(liveaction):
     # Use the user context from the parent action execution. Subtasks in a workflow
     # action can be invoked by a system user and so we want to use the user context
     # from the original workflow action.
-    parent_context = get_parent_context(liveaction)
+    parent_context = executions.get_parent_context(liveaction)
     if parent_context:
         parent_user = parent_context.get('user', None)
         if parent_user:
@@ -211,19 +211,6 @@ def request_cancellation(liveaction, requester):
     execution = ActionExecution.get(liveaction__id=str(liveaction.id))
 
     return (liveaction, execution)
-
-
-def get_parent_context(liveaction_db):
-    """
-    Returns context of the parent execution.
-
-    :return: If found the parent context else None.
-    :rtype: dict
-    """
-    context = getattr(liveaction_db, 'context', None)
-    if not context:
-        return None
-    return liveaction_db.context.get('parent', None)
 
 
 def _cleanup_liveaction(liveaction):
